Recreate all tables with non-nullable primary keys Fix the data type of a couple columns Improve logging during database creation and upgrade

+++ app/src/main/java/com/boardgamegeek/provider/BggDatabase.kt
@@ -533,7 +533,7 @@ class BggDatabase(private val context: Context?) : SQLiteOpenHelper(context, DAT
         .addColumn(Games.Columns.STATS_AVERAGE, ColumnType.REAL)
         .addColumn(Games.Columns.STATS_BAYES_AVERAGE, ColumnType.REAL)
         .addColumn(Games.Columns.STATS_STANDARD_DEVIATION, ColumnType.REAL)
-        .addColumn(Games.Columns.STATS_MEDIAN, ColumnType.INTEGER)
+        .addColumn(Games.Columns.STATS_MEDIAN, ColumnType.REAL)
         .addColumn(Games.Columns.STATS_NUMBER_OWNED, ColumnType.INTEGER)
         .addColumn(Games.Columns.STATS_NUMBER_TRADING, ColumnType.INTEGER)
         .addColumn(Games.Columns.STATS_NUMBER_WANTING, ColumnType.INTEGER)
